Feat(post): Finish edit user post funct

+++ src/module/posts/application/command/create-user-post/create-user-post.handler.ts
@@ -145,7 +145,9 @@ export class CreateUserPostHandler {
     } catch (err) {
       console.log(err)
       savedImages = await this.imageService.getImageByApplicableId(post.id)
-      await this.imageService.removeMultipleImages(savedImages!)
+      if (savedImages) {
+        await this.imageService.removeMultipleImages(savedImages!)
+      }
 
       if (
         err instanceof DomainError ||
